refactor: Socket 메시지 응답값 추가
- Service 분리
- 사용하지 않는 큐 제거
